+++ docs/index.md
@@ -38,9 +38,4 @@ Refer to [pyproject.toml](https://github.com/mozilla-ai/any-agent/blob/main/pypr
 
 ## For AI Systems
 
-This documentation is available in two AI-friendly formats:
-
-- **[llms.txt](/llms.txt)** - A structured overview with curated links to key documentation sections
 - **[llms-full.txt](/llms-full.txt)** - Complete documentation content concatenated into a single file
-
-Both files are dynamically generated at build time following the [llms.txt specification](https://llmstxt.org/).
